Added Main Folder
The plan eventually is to have a working application with a GUI that allows you to input a URL from reddit and the webscraper returns the data in a .xlsx format. I'm also considering connecting to a Google spreadsheet, and adding an API that connects to the Google sheet and adds to it. 

For now, though, this is the working code